Put openFile funciton in events.c
