+++ events.c
@@ -2,6 +2,7 @@
 
 #include <gtksourceview/gtksourcebuffer.h>
 
+#include <stdlib.h>
 
 #include "events.h"
 #include "graphics.h"
@@ -17,7 +18,7 @@ unsigned int play = 0;
 
 GtkTextMark* textMarker;
 
-
+int fileLength(FILE* f_input);
 
 gboolean events_mainLoop(gpointer user_data){
   /*https://developer.gnome.org/gtk-tutorial/stable/c1759.html*/
@@ -70,6 +71,37 @@ void events_init(GtkSourceBuffer* sourcebuffer)
                            TRUE);
 
   events_toggle();
+}
+
+void events_openFile(char* filename)
+{
+  FILE* f_input = NULL;
+  char *contents;
+  int length = 0;
+
+  f_input = fopen(filename, "r");
+  if((f_input = fopen(filename, "r"))){
+    length = fileLength(f_input);
+    contents = (char*)calloc(length, sizeof(char));
+    rewind(f_input);
+    fread(contents, sizeof(char), length, f_input);
+    fclose(f_input);
+    free(contents);  
+  }
+  else{
+    printf("Couldn't open file\n");
+  }
+}
+
+int fileLength(FILE* f_input)
+{
+  int count = 0;
+
+  while(!feof(f_input)){
+    getc(f_input);
+    count++;
+  }
+  return count;
 }
 
 void events_launchText(GtkSourceBuffer *sourcebuffer){
