+++ events.h
@@ -11,6 +11,7 @@ void events_start(void);
 void events_stop(void);
 void events_quitting(void);
 void events_launchText(GtkSourceBuffer *sourcebuffer);
+void events_openFile(char* filename);
 void events_buttonPress(GtkButton *);
 void events_init(GtkSourceBuffer* sourcebuffer);
